+++ demos/chinook/src/main/java/is/codion/framework/demos/chinook/testing/scenarios/UpdateTotals.java
@@ -63,5 +63,9 @@ public final class UpdateTotals extends AbstractUsageScenario<EntityConnectionPr
       connection.rollbackTransaction();
       throw e;
     }
+    catch (Exception e) {
+      connection.rollbackTransaction();
+      throw new RuntimeException(e);
+    }
   }
 }