+++ demos/employees/src/main/java/is/codion/framework/demos/employees/minimal/EmployeesMinimalApp.java
@@ -101,7 +101,7 @@ public final class EmployeesMinimalApp {
               Department.DEPTNO.define()
                       .column()
                       .caption("Department name")
-                      .searchColumn(true)
+                      .searchable(true)
                       .nullable(false)
                       .maximumLength(14),
              Department.LOC.define()
@@ -122,7 +122,7 @@ public final class EmployeesMinimalApp {
               Employee.ENAME.define()
                       .column()
                       .caption("Name")
-                      .searchColumn(true)
+                      .searchable(true)
                       .nullable(false)
                       .maximumLength(10),
               Employee.DEPTNO.define()