+++ demos/world/src/main/java/is/codion/framework/demos/world/domain/WorldImpl.java
@@ -70,7 +70,7 @@ public final class WorldImpl extends DefaultDomain implements World {
             City.NAME.define()
                     .column()
                     .caption("Name")
-                    .searchColumn(true)
+                    .searchable(true)
                     .nullable(false)
                     .maximumLength(35),
             City.COUNTRY_CODE.define()
@@ -124,7 +124,7 @@ public final class WorldImpl extends DefaultDomain implements World {
             Country.NAME.define()
                     .column()
                     .caption("Name")
-                    .searchColumn(true)
+                    .searchable(true)
                     .nullable(false)
                     .maximumLength(52),
             // tag::item[]
@@ -153,7 +153,7 @@ public final class WorldImpl extends DefaultDomain implements World {
             Country.INDEPYEAR_SEARCHABLE.define()
                     .column()
                     .expression("to_char(indepyear)")
-                    .searchColumn(true)
+                    .searchable(true)
                     .readOnly(true),
             Country.POPULATION.define()
                     .column()