+++ framework/db-core/src/test/java/is/codion/framework/db/TestDomain.java
@@ -57,7 +57,7 @@ public final class TestDomain extends DefaultDomain {
             Department.NAME.define()
                     .column()
                     .caption(Department.NAME.name())
-                    .searchColumn(true)
+                    .searchable(true)
                     .maximumLength(14)
                     .nullable(false),
             Department.LOCATION.define()
@@ -97,7 +97,7 @@ public final class TestDomain extends DefaultDomain {
             Employee.NAME.define()
                     .column()
                     .caption(Employee.NAME.name())
-                    .searchColumn(true)
+                    .searchable(true)
                     .name("ename")
                     .maximumLength(10)
                     .nullable(false),
@@ -111,7 +111,7 @@ public final class TestDomain extends DefaultDomain {
                     .column()
                     .items(asList(item("ANALYST"), item("CLERK"), item("MANAGER"), item("PRESIDENT"), item("SALESMAN")))
                     .caption(Employee.JOB.name())
-                    .searchColumn(true),
+                    .searchable(true),
             Employee.SALARY.define()
                     .column()
                     .caption(Employee.SALARY.name())