+++ framework/domain/src/test/java/is/codion/framework/domain/TestDomain.java
@@ -440,7 +440,7 @@ public final class TestDomain extends DefaultDomain {
             Department.NAME.define()
                     .column()
                     .caption(Department.NAME.name())
-                    .searchColumn(true)
+                    .searchable(true)
                     .maximumLength(14)
                     .nullable(false)
                     .beanProperty("name"),
@@ -520,7 +520,7 @@ public final class TestDomain extends DefaultDomain {
             Employee.NAME.define()
                     .column()
                     .caption(Employee.NAME.name())
-                    .searchColumn(true)
+                    .searchable(true)
                     .name("ename")
                     .maximumLength(10)
                     .nullable(false)
@@ -538,7 +538,7 @@ public final class TestDomain extends DefaultDomain {
                     .items(asList(item("ANALYST"), item("CLERK"),
                             item("MANAGER"), item("PRESIDENT"), item("SALESMAN")))
                     .caption(Employee.JOB.name())
-                    .searchColumn(true)
+                    .searchable(true)
                     .beanProperty("job"),
             Employee.SALARY.define()
                     .column()