+++ framework/json-db/src/test/java/is/codion/framework/json/TestDomain.java
@@ -98,7 +98,7 @@ public final class TestDomain extends DefaultDomain {
                     .updatable(true).nullable(false),
             Department.NAME.define()
                     .column()
-                    .searchColumn(true)
+                    .searchable(true)
                     .maximumLength(14)
                     .nullable(false),
             Department.LOCATION.define()
@@ -131,7 +131,7 @@ public final class TestDomain extends DefaultDomain {
                     .primaryKey(),
             Employee.NAME.define()
                     .column()
-                    .searchColumn(true).maximumLength(10).nullable(false),
+                    .searchable(true).maximumLength(10).nullable(false),
             Employee.DEPARTMENT.define()
                     .column()
                     .nullable(false),
@@ -140,7 +140,7 @@ public final class TestDomain extends DefaultDomain {
             Employee.JOB.define()
                     .column()
                     .items(asList(item("ANALYST"), item("CLERK"), item("MANAGER"), item("PRESIDENT"), item("SALESMAN")))
-                    .searchColumn(true),
+                    .searchable(true),
             Employee.SALARY.define()
                     .column()
                     .nullable(false).valueRange(1000, 10000).maximumFractionDigits(2),