+++ framework/model/src/main/java/is/codion/framework/model/EntitySearchModel.java
@@ -226,7 +226,7 @@ public interface EntitySearchModel {
    * @param entityType the type of the entity to search
    * @param connectionProvider the EntityConnectionProvider to use when performing the search
    * @return a new {@link EntitySearchModel.Builder} instance
-   * @see is.codion.framework.domain.entity.EntityDefinition.Columns#searchColumns()
+   * @see is.codion.framework.domain.entity.EntityDefinition.Columns#searchable()
    */
   static EntitySearchModel.Builder builder(EntityType entityType, EntityConnectionProvider connectionProvider) {
     return new DefaultEntitySearchModel.DefaultBuilder(entityType, connectionProvider);