+++ swing/framework-model-tools/src/test/java/is/codion/swing/framework/model/tools/loadtest/TestDomain.java
@@ -46,7 +46,7 @@ public final class TestDomain extends DefaultDomain {
             DEPARTMENT_NAME.define()
                     .column()
                     .caption(DEPARTMENT_NAME.name())
-                    .searchColumn(true)
+                    .searchable(true)
                     .maximumLength(14)
                     .nullable(false),
             DEPARTMENT_LOCATION.define()