+++ swing/framework-ui/src/main/java/is/codion/swing/framework/ui/component/EntitySearchField.java
@@ -765,9 +765,9 @@ public final class EntitySearchField extends HintTextField {
               .action(Control.control(table.searchField()::requestFocusInWindow))
               .enable(table);
       tableModel.columnModel().columns().forEach(this::configureColumn);
-      Collection<Column<String>> searchColumns = searchModel.columns();
-      tableModel.columnModel().setVisibleColumns(searchColumns.toArray(new Attribute[0]));
-      tableModel.sortModel().setSortOrder(searchColumns.iterator().next(), SortOrder.ASCENDING);
+      Collection<Column<String>> columns = searchModel.columns();
+      tableModel.columnModel().setVisibleColumns(columns.toArray(new Attribute[0]));
+      tableModel.sortModel().setSortOrder(columns.iterator().next(), SortOrder.ASCENDING);
       scrollPane = new JScrollPane(table);
       searchPanel.add(table.searchField(), BorderLayout.WEST);
       searchPanel.add(resultLimitLabel, BorderLayout.CENTER);